Move list into new 'todo list' component and create 'add todo form' component
