+++ src/App.js
@@ -1,35 +1,19 @@
 import React from "react";
+import TodoList from "./TodoList";
+import AddTodoForm from "./AddTodoForm";
 
-const todoList = [
-  {
-    id: 0,
-    title: "Eat 6 servings of vegetables",
-  },
-  {
-    id: 1,
-    title: "Drink 2 liters of water",
-  },
-  {
-    id: 2,
-    title: "Work out for 1 hour",
-  },
-];
-
-function App() {
+const App = () => {
   return (
     <div style={{ textAlign: "center" }}>
       <h1>Todo List</h1>
-      <ul>
-        {todoList.map(function (item) {
-          return (
-            <li key={item.id}>
-              <span>{item.title}</span>
-            </li>
-          );
-        })}
-      </ul>
+
+      <hr />
+
+      <TodoList />
+
+      <AddTodoForm />
     </div>
   );
-}
+};
 
 export default App;
